+++ internal/providersdk/provider.go
@@ -189,7 +189,6 @@ func Provider() *schema.Provider {
 			"junos_rstp":                                                 resourceRstp(),
 			"junos_rstp_interface":                                       resourceRstpInterface(),
 			"junos_security":                                             resourceSecurity(),
-			"junos_security_address_book":                                resourceSecurityAddressBook(),
 			"junos_security_dynamic_address_feed_server":                 resourceSecurityDynamicAddressFeedServer(),
 			"junos_security_dynamic_address_name":                        resourceSecurityDynamicAddressName(),
 			"junos_security_global_policy":                               resourceSecurityGlobalPolicy(),